requirements.txt
lz4==4.0.1
mss==6.1.0
numpy==1.23.0
Pillow==9.1.1
pygame==2.1.2
git+https://github.com/ra1nty/DXcam#egg=dxcam
